requirements.txt
attrs==19.3.0
certifi==2020.6.20
click==7.1.2
dash==1.16.3
dash-core-components==1.3.1
dash-html-components==1.0.1
dash-renderer==1.1.2
dash-table==4.4.1
dataclasses==0.6
fastapi==0.59.0
Flask==1.1.2
flask-compress==1.7.0
future==0.18.2
h11==0.9.0
httptools==0.1.1
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
MarkupSafe==1.1.1
more-itertools==8.4.0
mypy==0.782
mypy-extensions==0.4.3
numpy==1.19.0
packaging==20.4
pandas==1.0.5
plotly==4.11.0
pluggy==0.13.1
py==1.9.0
pydantic==1.6
pyparsing==2.4.7
pytest==5.4.3
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
retrying==1.3.3
requests==2.24.0
scikit-learn==0.23.1
scipy==1.5.1
six==1.15.0
starlette==0.13.4
threadpoolctl==2.1.0
typed-ast==1.4.1
typing-extensions==3.7.4.2
uvicorn==0.11.5
uvloop==0.14.0
wcwidth==0.2.5
websockets==8.1
Werkzeug==1.0.1
zipp==3.1.0
